+++ packages/react-native-updates/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+### [1.2.1-alpha.0](https://github.com/uplift-ltd/nexus/compare/@uplift-ltd/react-native-updates@1.2.0...@uplift-ltd/react-native-updates@1.2.1-alpha.0) (2022-07-27)
+
+**Note:** Version bump only for package @uplift-ltd/react-native-updates
+
+
+
+
+
 ## [1.2.0](https://github.com/uplift-ltd/nexus/compare/@uplift-ltd/react-native-updates@1.1.0...@uplift-ltd/react-native-updates@1.2.0) (2021-07-12)